requirements.txt
certifi==2021.10.8
Faker==12.3.0
kafka-python==2.0.2
python-dateutil==2.8.2
six==1.16.0
certifi==2021.10.8
Faker==12.3.0
kafka-python==2.0.2
mysql-connector-python==8.0.28
protobuf==3.19.4
python-dateutil==2.8.2
six==1.16.0
